src/backend/agentic-fleet/frontend/requirements.txt
chainlit>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
asyncio>=3.4.3
aiohttp>=3.9.1
fastapi>=0.105.0
uvicorn>=0.24.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
pydantic>=2.5.2
